experiment: whole C++ tree compiles for arm64-v8a
All 371 translation units in the Android build now compile cleanly. The work
splits into four kinds of change:

Architecture:
* core/simd.h - SSE/SSE2 fast paths keep working on ARM through sse2neon
  (vendored); agi/surface.cpp and agigl/gltexdef.cpp go through it
* agiworld/mmx.h, agiworld/meshlight.cpp, data7/mmx.cpp - the MMX vertex
  lighting path is x86-only; the existing scalar loop is the fallback, and
  UseMMX/HaveKNI stay false elsewhere
* agi/pipeline.cpp - x87 precision control is x86-only

Platform (POSIX implementations behind #ifdef _WIN32):
* core/wincompat.{h,cpp} - the Win32 file, directory and console calls the
  engine makes directly, over fds, dirent/fnmatch and ANSI colours; logging
  reaches logcat on Android
* stream/filestream.cpp - FileStream on fds, with a private mapping standing
  in for the FILE_MAP_COPY view
* memory/valloc.cpp - reserve/commit/decommit/release on mmap+mprotect
* data7/timer.cpp, data7/utimer.cpp - sched_yield, cpu relax, and a monotonic
  clock in place of RDTSC
* memory/allocator.cpp, arts7/sim.cpp - structured exception handling paths
  are Windows-only
* agi/screenshot.cpp - clipboard DIB export is Windows-only; the PNG is not

64-bit:
* agi/surface.h, agi/texdef.cpp - the paged texture header keeps 32-bit
  pointer fields on disk, so it is read through a packed form and copied
* data7/metatype.h/.cpp, mmbangers/banger.cpp, mmcity/inst.cpp - i32/u32
  overrides that only matched their isize/usize declarations on a 32-bit build

Non-dev build:
* arts7/cullmgr.cpp, arts7/midgets.cpp, mmcity/{cullcity,portal}.cpp,
  mmai/{aiMap,aiVehicleOpponent}.cpp - uses of debug pages, stats and widgets
  are now guarded the same way their declarations are

Excluded from the Android build: agisw, mmnetwork, mmwidget, mmaudio (all
DirectSound/EAX/MCI), the DirectInput force-feedback devices, mrkni.cpp,
and the Win32-only TUs.

+++ code/midtown/data7/metatype.cpp
@@ -168,7 +168,7 @@ struct FloatType final : MetaType
 {
 public:
     // ?Delete@FloatType@@UAEXPAXH@Z
-    ARTS_EXPORT void Delete(void* ptr, i32 len) override
+    ARTS_EXPORT void Delete(void* ptr, isize len) override
     {
         MetaDelete<f32>(ptr, len);
     }
@@ -180,7 +180,7 @@ public:
     }
 
     // ?New@FloatType@@UAEPAXH@Z
-    ARTS_EXPORT void* New(i32 count) override
+    ARTS_EXPORT void* New(isize count) override
     {
         return MetaNew<f32>(count);
     }
